rename ScryfallCard to ClientCard for clarity
