+++ frontend/src/Sale/SaleSearchCardList.tsx
@@ -3,13 +3,13 @@ import SearchIcon from '@material-ui/icons/Search';
 import React, { FC } from 'react';
 import Loading from '../ui/Loading';
 import Placeholder from '../ui/Placeholder';
-import { ScryfallCard } from '../utils/ScryfallCard';
+import { ClientCard } from '../utils/ClientCard';
 import SaleSearchCard from './SaleSearchCard';
 
 interface Props {
     loading: boolean;
     term: string;
-    cards: ScryfallCard[];
+    cards: ClientCard[];
 }
 
 const BrowseCardList: FC<Props> = ({ loading, term, cards }) => {
